move helper functions to `internal` module
